+++ axelor-business-project/src/main/java/com/axelor/apps/businessproject/service/ProjectBusinessServiceImpl.java
@@ -17,14 +17,14 @@
  */
 package com.axelor.apps.businessproject.service;
 
-import com.axelor.apps.account.db.InvoiceLine;
 import com.axelor.apps.base.db.Company;
 import com.axelor.apps.base.db.Partner;
-import com.axelor.apps.base.db.Product;
 import com.axelor.apps.businessproject.service.app.AppBusinessProjectService;
 import com.axelor.apps.project.db.Project;
 import com.axelor.apps.project.db.repo.ProjectRepository;
 import com.axelor.apps.project.service.ProjectServiceImpl;
+import com.axelor.apps.purchase.db.PurchaseOrder;
+import com.axelor.apps.purchase.service.PurchaseOrderService;
 import com.axelor.apps.sale.db.SaleOrder;
 import com.axelor.apps.sale.db.SaleOrderLine;
 import com.axelor.apps.sale.db.repo.SaleOrderRepository;
@@ -34,8 +34,6 @@ import com.axelor.exception.AxelorException;
 import com.axelor.inject.Beans;
 import com.google.inject.Inject;
 import com.google.inject.persist.Transactional;
-import java.math.BigDecimal;
-import java.util.List;
 
 public class ProjectBusinessServiceImpl extends ProjectServiceImpl
     implements ProjectBusinessService {
@@ -56,7 +54,7 @@ public class ProjectBusinessServiceImpl extends ProjectServiceImpl
     order.setClientPartner(project.getClientPartner());
     return Beans.get(SaleOrderRepository.class).save(order);
   }
-
+  
   /**
    * Generate project form SaleOrder and set bi-directional.
    *
@@ -72,7 +70,6 @@ public class ProjectBusinessServiceImpl extends ProjectServiceImpl
             saleOrder.getSalemanUser(),
             saleOrder.getCompany(),
             saleOrder.getClientPartner());
-    project.setSaleOrder(saleOrder);
     saleOrder.setProject(project);
     return project;
   }
@@ -92,15 +89,6 @@ public class ProjectBusinessServiceImpl extends ProjectServiceImpl
     if (parentProject != null) {
       project.setProjInvTypeSelect(parentProject.getProjInvTypeSelect());
     }
-    Product product =
-        appBusinessProjectService.getAppBusinessProject().getProductInvoicingProject();
-    if (product != null) {
-      project.setProduct(product);
-      project.setQty(BigDecimal.ONE);
-      project.setPrice(product.getPurchasePrice());
-      project.setUnit(product.getUnit());
-      project.setExTaxTotal(product.getPurchasePrice());
-    }
     return project;
   }
 
@@ -113,32 +101,8 @@ public class ProjectBusinessServiceImpl extends ProjectServiceImpl
             saleOrderLine.getSaleOrder().getSalemanUser(),
             parent.getCompany(),
             parent.getClientPartner());
-    project.setProduct(saleOrderLine.getProduct());
-    project.setQty(saleOrderLine.getQty());
-    project.setPrice(saleOrderLine.getPrice());
-    project.setUnit(saleOrderLine.getUnit());
-    project.setExTaxTotal(saleOrderLine.getCompanyExTaxTotal());
     project.setProjectTypeSelect(ProjectRepository.TYPE_PHASE);
     saleOrderLine.setProject(project);
     return project;
-  }
-
-  /**
-   * Manages invoice lines for project dashlets
-   *
-   * @param invoiceLine InvoiceLine to add or remove
-   * @param project Project to add or remove the invoice line
-   */
-  @Override
-  @Transactional(rollbackOn = {AxelorException.class, Exception.class})
-  public void manageInvoiceLine(InvoiceLine invoiceLine, Project project) {
-    List<InvoiceLine> invoiceLines = project.getInvoiceLineList();
-    if (invoiceLines.contains(invoiceLine)) {
-      project.removeInvoiceLineListItem(invoiceLine);
-    } else {
-      project.addInvoiceLineListItem(invoiceLine);
-    }
-
-    projectRepo.save(project);
   }
 }