+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/db/repo/TeamTaskHRRepository.java
@@ -33,9 +33,11 @@ public class TeamTaskHRRepository extends TeamTaskProjectRepository {
     super.save(teamTask);
 
     teamTask.setTotalPlannedHrs(projectPlanningTimeService.getTaskPlannedHrs(teamTask));
+    teamTask.setTotalRealHrs(projectPlanningTimeService.getTaskRealHrs(teamTask));
 
     Project project = teamTask.getProject();
     project.setTotalPlannedHrs(projectPlanningTimeService.getProjectPlannedHrs(project));
+    project.setTotalRealHrs(projectPlanningTimeService.getProjectRealHrs(project));
 
     return teamTask;
   }
@@ -47,5 +49,6 @@ public class TeamTaskHRRepository extends TeamTaskProjectRepository {
     super.remove(teamTask);
 
     project.setTotalPlannedHrs(projectPlanningTimeService.getProjectPlannedHrs(project));
+    project.setTotalRealHrs(projectPlanningTimeService.getProjectRealHrs(project));
   }
 }